+++ src/main/java/components/flipflops/FlipFlopD.java
@@ -10,7 +10,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 
 public class FlipFlopD extends FlipFlop{
-    public FlipFlopD(double x, double y, boolean startLife, XYChart.Series<Integer, String> series, AtomicInteger chartMillisCounter){
+    public FlipFlopD(double x, double y, boolean startLife, XYChart.Series<Long, String> series, Long chartMillisCounter){
         super(x, y, startLife, series, chartMillisCounter);
 
         name = Names.flipFlopD;
