+++ src/main/java/components/flipflops/FlipFlopJK.java
@@ -16,7 +16,7 @@ public class FlipFlopJK extends FlipFlop{
     private ArrayList<Line> arrayListLinesInputK = new ArrayList<>();
     private Point pointInputK;
 
-    public FlipFlopJK(double x, double y, boolean startLife, XYChart.Series<Integer, String> series, AtomicInteger chartMillisCounter){
+    public FlipFlopJK(double x, double y, boolean startLife, XYChart.Series<Long, String> series, Long chartMillisCounter){
         super(x, y, startLife, series, chartMillisCounter);
 
         name = Names.flipFlopJK;
