+++ src/main/java/components/flipflops/FlipFlopT.java
@@ -9,7 +9,7 @@ import javafx.scene.image.ImageView;
 import java.util.concurrent.atomic.AtomicInteger;
 
 public class FlipFlopT extends FlipFlop {
-    public FlipFlopT(double x, double y, boolean startLife, XYChart.Series<Integer, String> series, AtomicInteger chartMillisCounter){
+    public FlipFlopT(double x, double y, boolean startLife, XYChart.Series<Long, String> series, Long chartMillisCounter){
         super(x, y, startLife, series, chartMillisCounter);
 
         name = Names.flipFlopT;
