+++ src/main/java/components/gates/xnor/Xnor.java
@@ -9,7 +9,7 @@ import javafx.scene.chart.XYChart;
 import java.util.concurrent.atomic.AtomicInteger;
 
 public abstract class Xnor extends Gate {
-    public Xnor(double x, double y, boolean startLife, XYChart.Series<Integer, String> series, AtomicInteger chartMillisCounter){
+    public Xnor(double x, double y, boolean startLife, XYChart.Series<Long, String> series, Long chartMillisCounter){
         super(x, y, startLife, series, chartMillisCounter);
     }
